Rename test files for stagin heap (#209)
* D3D11 buffers use staging type if purely for copying

* Split staging heap tests into 2 files

* Rename test-upload->test-staging-heap

* Split staging heap tests into 2 files

* Rename test-upload->test-staging-heap
